Housekeep JSON in verify
