+++ internal/verify/diff.go
@@ -1,18 +1,16 @@
 package verify
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
-	"log"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
 	tftags "github.com/hashicorp/terraform-provider-aws/internal/tags"
-	awspolicy "github.com/jen20/awspolicyequivalence"
 )
+
+// Find JSON diff functions in the json.go file.
 
 // SetTagsDiff sets the new plan difference with the result of
 // merging resource tags on to those defined at the provider-level;
@@ -55,15 +53,6 @@ func SetTagsDiff(_ context.Context, diff *schema.ResourceDiff, meta interface{})
 	return nil
 }
 
-func SuppressEquivalentPolicyDiffs(k, old, new string, d *schema.ResourceData) bool {
-	equivalent, err := awspolicy.PoliciesAreEquivalent(old, new)
-	if err != nil {
-		return false
-	}
-
-	return equivalent
-}
-
 // SuppressEquivalentTypeStringBoolean provides custom difference suppression for TypeString booleans
 // Some arguments require three values: true, false, and "" (unspecified), but
 // confusing behavior exists when converting bare true/false values with state.
@@ -83,38 +72,6 @@ func SuppressEquivalentTypeStringBoolean(k, old, new string, d *schema.ResourceD
 //  * The operator's configuration omits the optional configuration block
 func SuppressMissingOptionalConfigurationBlock(k, old, new string, d *schema.ResourceData) bool {
 	return old == "1" && new == "0"
-}
-
-func SuppressEquivalentJSONDiffs(k, old, new string, d *schema.ResourceData) bool {
-	ob := bytes.NewBufferString("")
-	if err := json.Compact(ob, []byte(old)); err != nil {
-		return false
-	}
-
-	nb := bytes.NewBufferString("")
-	if err := json.Compact(nb, []byte(new)); err != nil {
-		return false
-	}
-
-	return JSONBytesEqual(ob.Bytes(), nb.Bytes())
-}
-
-func SuppressEquivalentJSONOrYAMLDiffs(k, old, new string, d *schema.ResourceData) bool {
-	normalizedOld, err := NormalizeJSONOrYAMLString(old)
-
-	if err != nil {
-		log.Printf("[WARN] Unable to normalize Terraform state CloudFormation template body: %s", err)
-		return false
-	}
-
-	normalizedNew, err := NormalizeJSONOrYAMLString(new)
-
-	if err != nil {
-		log.Printf("[WARN] Unable to normalize Terraform configuration CloudFormation template body: %s", err)
-		return false
-	}
-
-	return normalizedOld == normalizedNew
 }
 
 // DiffStringMaps returns the set of keys and values that must be created, the set of keys
