+++ internal/verify/diff_test.go
@@ -3,33 +3,7 @@ package verify
 import (
 	"reflect"
 	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
-
-func TestSuppressEquivalentJSONDiffsWhitespaceAndNoWhitespace(t *testing.T) {
-	d := new(schema.ResourceData)
-
-	noWhitespace := `{"test":"test"}`
-	whitespace := `
-{
-  "test": "test"
-}`
-
-	if !SuppressEquivalentJSONDiffs("", noWhitespace, whitespace, d) {
-		t.Errorf("Expected SuppressEquivalentJSONDiffs to return true for %s == %s", noWhitespace, whitespace)
-	}
-
-	noWhitespaceDiff := `{"test":"test"}`
-	whitespaceDiff := `
-{
-  "test": "tested"
-}`
-
-	if SuppressEquivalentJSONDiffs("", noWhitespaceDiff, whitespaceDiff, d) {
-		t.Errorf("Expected SuppressEquivalentJSONDiffs to return false for %s == %s", noWhitespaceDiff, whitespaceDiff)
-	}
-}
 
 func TestSuppressEquivalentTypeStringBoolean(t *testing.T) {
 	testCases := []struct {
@@ -68,200 +42,6 @@ func TestSuppressEquivalentTypeStringBoolean(t *testing.T) {
 
 		if !tc.equivalent && value {
 			t.Fatalf("expected test case %d to not be equivalent", i)
-		}
-	}
-}
-
-func TestSuppressEquivalentJSONOrYAMLDiffs(t *testing.T) {
-	testCases := []struct {
-		description string
-		equivalent  bool
-		old         string
-		new         string
-	}{
-		{
-			description: `JSON no change`,
-			equivalent:  true,
-			old: `
-{
-   "Resources": {
-      "TestVpc": {
-         "Type": "AWS::EC2::VPC",
-         "Properties": {
-            "CidrBlock": "10.0.0.0/16"
-         }
-      }
-   },
-   "Outputs": {
-      "TestVpcID": {
-         "Value": { "Ref" : "TestVpc" }
-      }
-   }
-}
-`,
-			new: `
-{
-   "Resources": {
-      "TestVpc": {
-         "Type": "AWS::EC2::VPC",
-         "Properties": {
-            "CidrBlock": "10.0.0.0/16"
-         }
-      }
-   },
-   "Outputs": {
-      "TestVpcID": {
-         "Value": { "Ref" : "TestVpc" }
-      }
-   }
-}
-`,
-		},
-		{
-			description: `JSON whitespace`,
-			equivalent:  true,
-			old:         `{"Resources":{"TestVpc":{"Type":"AWS::EC2::VPC","Properties":{"CidrBlock":"10.0.0.0/16"}}},"Outputs":{"TestVpcID":{"Value":{"Ref":"TestVpc"}}}}`,
-			new: `
-{
-   "Resources": {
-      "TestVpc": {
-         "Type": "AWS::EC2::VPC",
-         "Properties": {
-            "CidrBlock": "10.0.0.0/16"
-         }
-      }
-   },
-   "Outputs": {
-      "TestVpcID": {
-         "Value": { "Ref" : "TestVpc" }
-      }
-   }
-}
-`,
-		},
-		{
-			description: `JSON change`,
-			equivalent:  false,
-			old: `
-{
-   "Resources": {
-      "TestVpc": {
-         "Type": "AWS::EC2::VPC",
-         "Properties": {
-            "CidrBlock": "10.0.0.0/16"
-         }
-      }
-   },
-   "Outputs": {
-      "TestVpcID": {
-         "Value": { "Ref" : "TestVpc" }
-      }
-   }
-}
-`,
-			new: `
-{
-   "Resources": {
-      "TestVpc": {
-         "Type": "AWS::EC2::VPC",
-         "Properties": {
-            "CidrBlock": "172.16.0.0/16"
-         }
-      }
-   },
-   "Outputs": {
-      "TestVpcID": {
-         "Value": { "Ref" : "TestVpc" }
-      }
-   }
-}
-`,
-		},
-		{
-			description: `YAML no change`,
-			equivalent:  true,
-			old: `
-Resources:
-  TestVpc:
-    Type: AWS::EC2::VPC
-    Properties:
-      CidrBlock: 10.0.0.0/16
-Outputs:
-  TestVpcID:
-    Value: !Ref TestVpc
-`,
-			new: `
-Resources:
-  TestVpc:
-    Type: AWS::EC2::VPC
-    Properties:
-      CidrBlock: 10.0.0.0/16
-Outputs:
-  TestVpcID:
-    Value: !Ref TestVpc
-`,
-		},
-		{
-			description: `YAML whitespace`,
-			equivalent:  false,
-			old: `
-Resources:
-  TestVpc:
-    Type: AWS::EC2::VPC
-    Properties:
-      CidrBlock: 10.0.0.0/16
-
-Outputs:
-  TestVpcID:
-    Value: !Ref TestVpc
-
-`,
-			new: `
-Resources:
-  TestVpc:
-    Type: AWS::EC2::VPC
-    Properties:
-      CidrBlock: 10.0.0.0/16
-Outputs:
-  TestVpcID:
-    Value: !Ref TestVpc
-`,
-		},
-		{
-			description: `YAML change`,
-			equivalent:  false,
-			old: `
-Resources:
-  TestVpc:
-    Type: AWS::EC2::VPC
-    Properties:
-      CidrBlock: 172.16.0.0/16
-Outputs:
-  TestVpcID:
-    Value: !Ref TestVpc
-`,
-			new: `
-Resources:
-  TestVpc:
-    Type: AWS::EC2::VPC
-    Properties:
-      CidrBlock: 10.0.0.0/16
-Outputs:
-  TestVpcID:
-    Value: !Ref TestVpc
-`,
-		},
-	}
-
-	for _, tc := range testCases {
-		value := SuppressEquivalentJSONOrYAMLDiffs("test_property", tc.old, tc.new, nil)
-
-		if tc.equivalent && !value {
-			t.Fatalf("expected test case (%s) to be equivalent", tc.description)
-		}
-
-		if !tc.equivalent && value {
-			t.Fatalf("expected test case (%s) to not be equivalent", tc.description)
 		}
 	}
 }
